requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
pydantic-settings
python-dotenv
pytest
httpx
